+++ CHANGELOG.md
@@ -12,12 +12,57 @@ All the notable changes to this project will be documented in this file.
 ### Removed
 - 
 
-## 1.1.0
+## [1.7.0]
+
+### Added
+- Contact widget
+- PostFilter widget
+- Woocart widget
+- Post and Page default functions
+
+## [1.6.2]
+
+### Fixed
+- Sourcemap in js file
+
+## [1.6.0]
+
+### Added
+- Template self update from GitHub
+
+### Fixed
+- Other bug fixes
+
+## [1.5.0]
+
+### Added
+- Template class files
+
+### Fixed
+- Other bug fixes
+
+## [1.4.0]
+
+### Fixed
+- Fixed content layout styling
+- Other bug fixes
+
+## [1.3.0]
+
+### Fixed
+- Fixed navigation bug
+
+## [1.2.0]
+
+### Fixed
+- Fixed navigation bug
+
+## [1.1.0]
 
 ### Fixed
 - Fixed layout styling
 
-## 1.0.0
+## [1.0.0]
 
 ### Added
 - Default template parts and files